requirements.txt
Appium-Python-Client==5.2.4
pytest==8.4.2
allure-pytest==2.15.0
requests==2.31.0
responses==0.25.3
